# doc/apidoc/schema_swagger_json.json
{
  "swagger": "2.0",
  "info": {
    "title": "Myapp (params in:body)",
    "description": "\n<p>Another API description</p>\n",
    "version": "1.0",
    "x-copyright": null
  },
  "basePath": "",
  "consumes": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "consumes": [
          "application/json"
        ],
        "operationId": "get_users",
        "parameters": [

        ],
        "responses": {
          "200": {
            "description": "List of users",
            "schema": {
              "type": "object",
              "properties": {
                "users": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "name"
                      }
                    },
                    "additionalProperties": false,
                    "required": [
                      "name"
                    ]
                  }
                }
              },
              "additionalProperties": false,
              "required": [
                "users"
              ]
            }
          }
        },
        "description": ""
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "consumes": [
          "application/json"
        ],
        "operationId": "get_users_id",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The User",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        },
        "description": ""
      }
    }
  },
  "definitions": {
    "user": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "name"
        }
      },
      "additionalProperties": false,
      "required": [
        "name"
      ]
    }
  },
  "tags": [

  ],
  "host": "localhost:3000"
}